25-11-2022-1

+++ api-fe/src/app/components/usuario/usuario-edit/usuario-edit.component.ts
@@ -1,3 +1,4 @@
+import { Telefone } from './../../../model/telefone';
 import { UsuarioService } from 'src/app/service/usuario.service';
 import { User } from 'src/app/model/user';
 import { Component, Inject, OnInit } from '@angular/core';
@@ -14,7 +15,8 @@ export class UsuarioEditComponent implements OnInit {
   tab1: String = "active"
   tab2: String = ""
   tab3: String = ""
-  usuario = new User()
+  cdUsuario = new User()
+  telefone = new Telefone()
 
   constructor(public dialogRef: MatDialogRef<UsuarioEditComponent>, @Inject(MAT_DIALOG_DATA) private modalData: any, private spinner: NgxSpinnerService, private usuarioService: UsuarioService) { }
 
@@ -25,13 +27,13 @@ export class UsuarioEditComponent implements OnInit {
   getUsuario(id: Number) {
     if (id != 0) {
       this.usuarioService.getUsuarioId(id).subscribe((resposta) => {
-        this.usuario = resposta
+        this.cdUsuario = resposta
       })
     }
   }
 
-  salvar() {
-    this.usuarioService.salvaUsuario(this.usuario).subscribe((resposta) => {
+  salvarUsuario() {
+    this.usuarioService.salvaUsuario(this.cdUsuario).subscribe((resposta) => {
       this.fecharCad()
       const Toast = Swal.mixin({
         toast: true,
@@ -52,27 +54,40 @@ export class UsuarioEditComponent implements OnInit {
     })
   }
 
-  deleteTelefone(id: Number) {
-    Swal.fire({
-      title: "Deletar item",
-      text: "Deseja mesmo remover este item?",
-      icon: 'warning',
-      showCancelButton: true,
-      allowOutsideClick: false,
-      width: "740px",
-      confirmButtonColor: '#3085d6',
-      cancelButtonColor: '#d33',
-      confirmButtonText: 'Sim, pode continuar!',
-      cancelButtonText: 'Não, cancelar!'
-    }).then((result) => {
-      if (result.isConfirmed) {
-        this.spinner.show();
-        this.usuarioService.deleteTelefone(id).subscribe((resposta) => {
-          this.getUsuario(this.usuario.id) // retorna o usuario com os dados atualizados
-          this.spinner.hide()
-        });
-      }
-    })
+  salvarTelefone() {
+    if (this.cdUsuario.listTelefones === undefined) {
+      this.cdUsuario.listTelefones = new Array<Telefone>()
+    }
+
+    this.cdUsuario.listTelefones.push(this.telefone)
+    this.telefone = new Telefone()
+  }
+
+  deleteTelefone(id: Number, i: number) {
+    if (id > 0) {
+      Swal.fire({
+        title: "Deletar item",
+        text: "Deseja mesmo remover este item?",
+        icon: 'warning',
+        showCancelButton: true,
+        allowOutsideClick: false,
+        width: "740px",
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Sim, pode continuar!',
+        cancelButtonText: 'Não, cancelar!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.spinner.show();
+          this.usuarioService.deleteTelefone(id).subscribe((resposta) => {
+            this.getUsuario(this.cdUsuario.id) // retorna o usuario com os dados atualizados
+            this.spinner.hide()
+          });
+        }
+      })
+    } else {
+      this.cdUsuario.listTelefones.splice(i, 1)
+    }
   }
 
   fecharCad() {
